Support facets in the url that don't exist yet
For example, the page first loads with a domain=Vivaldi facet
in the URL. But there is no such domain, because domains are
not loaded for another .1 second. So the facet is in the URL
but not the search bar. With this fix, when domains are
loaded and the facet JSON is updated, the code that
regenerates the facets also rereads from the URL.

This is a revision of my little hack in bootstrap_magic_search.js
to handle options coming in late. It actually feels less
like a hack now, and more like something you might want
to pull in and use.